quick and dirty fix
A quick and dirty fix for the selection highlight problem.  I still need to
figure out the exact cause of what's happening, but this patch should at least
mask the problem.  All it does is make it so that gaps in selection are not
highlighted (which actually matches the behavior of firefox, opera and IE).