fix for caret position in empty table cell
Don't include the intrinsic padding when computing the caret's y-coordinate in
an empty table cell.  Note: due to recent changes in WebKit, this patch will
not apply after webkit's rev 109818.

To apply this rev after 109818, change the 'false' argument to
'ExcludeIntrinsicPadding'.

To apply this rev after 113823, call 'computedCSSPaddingTop' instead of
'paddingTop'.