only include \n if next node is text
DRQS 31922188
Don't include the \n for empty paragraphs if the next node is not a text node.
This is so that moveParagraphWithClones will not include the next node (which
happens when the end position is canonicalized).

This is one of those changes that should NOT be upstreamed because it involves
the editing component of WebKit, which appears to be very brittle and we don't
know what ramifications it will have in other areas.  Based on discussion with
rniwa, all this code will/should be rewritten anyway based on
http://dvcs.w3.org/hg/editing/raw-file/tip/editing.html