dont clone outerNode if blockElement is descendent
DRQS 32165963
This case happens when an image is indented twice within an editable body,
where the splitNodes logic doesn't work quite right.  Don't upstream this
change because the entire html editing code is going to be rewritten and the
webkit devs don't want to make too many changes to this version.