add dom properties to control marker colors
DRQS 32415776
Adding this as a DOM node property for now instead of a CSS property, in order
to make a small change as possible.  Right now, only alpha=0 is supported to
hide the marker.  If alpha!=0, then the marker will be opaque.  This is due to
some strange issues I saw when trying to paint semi-transparent markers.