use getAttribute instead of attributeItem
This is because the getAttributeItemIndex and attributeItem methods are not
available in our branch of chromium.  Once we update, we would ideally want to
go back to use getAttributeItemIndex and attributeItem because getAttribute may
be doing other things like updateStyleAttribute.  I'm not sure if these have
any side-effects, I hope not.