fix the insertion point for next siblings
DRQS 32418391
Fix the logic that determines the insertion point for next siblings whenever
the start node is different from the end node.  We may want to upstream this
because the previous logic in WebKit was both wrong and also non-obvious.

The new logic is as follows: Before we start the while loop, 'lastNode' will
point to the clone of startNode.  We need to move lastNode up the tree the same
number of times that we need to move the startNode up in order for startNode
and nextSibling to have the same parent (i.e. to make them truly siblings).