transform destRect when computing resampling mode
DRQS 32353761
When computing the resampling mode, we should transform the destRect using the
current matrix in the canvas because this is the actual rect that the image
will be drawn against.
Also in this change, setFilterBitmap to true only if we're using linear
resampling (this change was already done upstream on may 15th).