fix flexbox issue when overrideHeight is set
DRQS 33177416
When flex direction is "row", flexbox uses overrideHeight to set the height of
it's children to match the height of the flexbox.

The logic that computes available free space does not subtract
preferredMainAxisExtent when overrideHeight is set, so the flexbox assumes that
it has its full height available for flexing.