Allow numeric values in -webkit-column-span
DRQS 33399541
This just expands the internal structures to support holding numeric values for
columnSpan, instead of just an enum of One and All.
There is no real logic change yet in the layout code, but we can start playing
around with that, now that the internal structures can actually hold the
values.