Build v8 to support dll mode
When BUILDING_V8_SHARED is defined, the V8EXPORT macro will resolve to
__declspec(dllexport), otherwise it resolves to __declspec(dllimport).

We define BUILDING_V8_SHARED in v8/src/v8.h, which is only included from within
the V8 build.  This means any external code will automatically start using
dllimport without needing to change their build configuration.

This approach is taken as opposed to modifying the vcproj files because the
vcproj files appear to be generated on a single line, which means we will
constantly get conflicts when we merge this change over.