Unite unscaled clipRect with scaled clipRect
DRQS 32861389
The previous logic just returned the scaled clipRect, which is to handle the
case where the transform makes the clipRect bigger.  However, when the
transform makes the clipRect smaller, then we should prefer to return the
original (bigger, unscaled) clipRect.

This patch handles both situations by simply uniting the scaled and unscaled
clipRects.

Some background research on this (sha's are from the webkit repo on github):
* transforming was introduced in 5baf4a9f7dcf93192e4ceca4be6fe29310229298
  (svn rev 51175)
* 3c8e821d67c7d1cb8cd858d9fb19a6526bc9cb10 (svn rev 55597) is interesting
  because it only seems to return "this" if we are compositing, but not
  transforming.  This leads to the discrepancy in behavior when CSS "overflow"
  is used, but that turned out to be a red-herring.
* 7fb19f23727bcbd4efd8692df8ce0be8f95ecea2 (svn rev 17002) was the original
  addition of childrenClipRect and selfClipRect, and this implementation did
  not transform the rect.