Use nextChild's logicalTop as spanningHeaderHeight
DRQS 33399541
The logicalHeight() of a block does not include the margins from its children.
As a result, if the spanning header contained child nodes with margins, those
margins were not taken into account when setting the spanning header's height.

With this change, the spanning header's height is made equivalent to the
logicalTop of the next child in the parent's layoutBlockChildren (i.e. the
nextSiblingBox() of the spanning header).  This logicalTop does take into
account the margins of any nodes within the spanning header.