Move outdentWithBR fix to outdentParagraph
https://bugs.webkit.org/show_bug.cgi?id=92130
This is likely to be closer to the fix that will go upstream.  Instead of
making the change in SplitElementCommand, we make the change in
IndentOutdentCommand::outdentParagraph.  This limits the scope of the change
(it was causing other test cases to fail).

Also, introduce a fix for the case where there is an inline element right
before the blockquote.  We insert a <br> element so that the moved paragraph
will not be on the same line as the previous inline.