Inherit zoom from owner document
DRQS 34484409
If the element has the html tag, and the document has an owner element, then
let the element inherit the effectiveZoom from the renderer of the owner
element.

Also, when updating the style for an iframe element, if the content's effective
zoom doesn't seem to have the owner's effective zoom into account, then make
the content recalc its style.