Set spanningHeaderHeight before next box's layout
DRQS 33399541
The previous logic set spanningHeaderHeight after laying out the next child
after the spanning header, so that we could get the logicalTop of the next
child.  However, this makes the layout of *that* child invalid if it spans
columns.

With this change, we move the logic that sets the spanningHeaderHeight to
happen before that next child is laid out.  Additionally, if the logicalTop of
the child changes after clearing floats, then the spanning height is set again.
Also, the child is marked for relayout if it has spanned multiple columns, so
that it's lineboxes are repaginated.