Support table cells in multicolumn layout
DRQS 34079255
Make table cells unsplittable for pagination.  Add a pagination strut field for
each row, and during the layoutRows method in RenderTableSection, calculate the
remaining space for the row and push the row down if there isn't enough space
for it.

Note that there is a lot of cleanup necessary, but this should work for us
right now.  Pending work include:
* remove the 'goto'
* remove the 'shez*' static functions that were copied from RenderBlock
* ensure that the row's pagination strut is taken into account whenever
  necessary when m_rowPos is referenced