Don't force relayout if percent-height is same
DRQS 36633185
The new logic is enabled by adding "bbcr_dont_relayout_if_percent_same" class
to the element with a percent-height.

In the previous logic, any element with a percent-height would be forced to
relayout when its parent was laid out.  When the new logic is enabled, we will
only force a layout if the computed logical height is different from the
existing logical height.