Prevent blocks laid out under spanning header
DRQS 36821222
When laying out RenderBlocks, there are some scenarios where the logicalTop of
the RenderBlock's child ends up between the top of the column and bottom of the
spanning header.  The result is that child appears "beneath" the spanning
header.

With this fix, the adjustLogicalTopForSpanningHeader tries to detect this
situation and corrects it by adjusting the logicalTop to be just below the
spanning header.  This function is called from various points in the code where
block layout is happening.