Handle multiple paragraphs on the same list item
DRQS 36805707
The outdentRegion function moves the entire list item, even if the region
provided is a single paragraph within that list item.  This was causing our
startOfNextParagraph position to become invalid, which caused the
loop-terminating condition to never be true.

To fix this, we detect when the current paragraph and the next paragraph are on
the same list item.  If so, we advance the startOfNextParagraph pointer
*before* calling outdentRegion.