Apply alignment to non-inside list markers
DRQS 37088493
https://bugs.webkit.org/show_bug.cgi?id=21709

Previously, alignment was applied to list markers only if list-position-style
was "inside".  Now, we do it for outside list markers too.  There are two
changes needed to make this work:

* InlineFlowBox::placeBoxesInInlineDirection now places list marker boxes even
  when they are not inside.  This is where the placement based on alignment
  happens.
* RenderListItem::positionListMarker no longer adjusts the position of the list
  marker (otherwise, it undoes the placement from placeBoxesInInlineDirection).

This change is enabled by a global flag "g_enableListMarkerFix_DRQS37088493",
which is there purely as a precaution so that we can disable this change
externally if we encounter any problems with it.