Set RenderText.width = 0 if it only contains a new line
Keep the RenderText's minimum/maximum text length to 0 if it only
contains a single new-line character. This is necessary for
RenderBlock::LineBreaker::nextLineBreak to accurately calculate
the minimum width of a line with a new line node at the end. Since
it's a breakable character, we do not want to allocate some pixels
(usually 4) if it's going to be the last character in the line.
Doing so will either push the character offscreen or move the
character to the next line, which will look very strange.