DRQS 32427286
Add an optional feature for the image resource, which allows to cancel
loading of that resource if no more clients are attached to it. This
feature is disabled by default.

Add a callback to notify a frame of html-image elements being inserted and
removed from DOM. This notification will allow to track image elements in
higher level and enable the fore-mentioned feature to cancel loading of
the image resource when necessary.