set the callback using a separate function
Prior to this change, the constructor had an argument for the
'all message processed' handler. If this was passed during the
construction of TestShellDevtoolsClient, it would get called
every time a batch of messages were processed by devtools.js.
Now, we have a function setAllMessageProcessedHandler that does
the same thing except it doesn't have to happen during the
construction of the object. This allows the handler to be reset
back to NULL once we are no longer interested in receiving
the notification.